# docs/openapi/knowledge.openapi.yaml
openapi: 3.1.0
info:
  title: knowledge
  version: 1.0.0
  description: knowledge 控制面、知识生产、发布、检索与 Agent Run API。
paths:
...
  /kbs/{kb_id}/documents:
    get:
      tags:
      - documents
      summary: List Documents
      operationId: list_documents_kbs__kb_id__documents_get
      security:
      - HTTPBearer: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: integer
          title: Kb Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentRead'
                title: Response List Documents Kbs  Kb Id  Documents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /kbs/{kb_id}/documents/{doc_id}:
    get:
      tags:
      - documents
      summary: Get Document
      operationId: get_document_kbs__kb_id__documents__doc_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: integer
          title: Kb Id
      - name: doc_id
        in: path
        required: true
        schema:
          type: integer
          title: Doc Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetailRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - documents
      summary: Delete Document
      operationId: delete_document_kbs__kb_id__documents__doc_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: integer
          title: Kb Id
      - name: doc_id
        in: path
        required: true
        schema:
          type: integer
          title: Doc Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /ops/overview:
    get:
      tags:
      - ops
      summary: Overview
      operationId: overview_ops_overview_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsOverviewResponse'
      security:
      - HTTPBearer: []
...
  /ops/workers:
    get:
      tags:
      - ops
      summary: Workers
      operationId: workers_ops_workers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WorkerStatusRead'
                type: array
                title: Response Workers Ops Workers Get
      security:
      - HTTPBearer: []
  /ops/tasks/failures:
    get:
      tags:
      - ops
      summary: Recent Task Failures
      operationId: recent_task_failures_ops_tasks_failures_get
      security:
      - HTTPBearer: []
      parameters:
      - name: trace_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskFailureRead'
                title: Response Recent Task Failures Ops Tasks Failures Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    DocumentChunkRead:
      properties:
        id:
          type: integer
          title: Id
        chunk_index:
          type: integer
          title: Chunk Index
        text:
          type: string
          title: Text
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
        created_at:
          type: string
          format: date-time
          title: Created At
        embedding_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Embedding Model
        index_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Index Status
      type: object
      required:
      - id
      - chunk_index
      - text
      - created_at
      title: DocumentChunkRead
    DocumentDetailRead:
      properties:
        id:
          type: integer
          title: Id
        source_path:
          type: string
          title: Source Path
        source_file_name:
          type: string
          title: Source File Name
        file_type:
          type: string
          title: File Type
        source_kind:
          type: string
          title: Source Kind
        parse_status:
          type: string
          title: Parse Status
        chunk_count:
          type: integer
          title: Chunk Count
          default: 0
        last_indexed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Indexed At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        source_etag_or_mtime:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Etag Or Mtime
        chunks:
          items:
            $ref: '#/components/schemas/DocumentChunkRead'
          type: array
          title: Chunks
      type: object
      required:
      - id
      - source_path
      - source_file_name
      - file_type
      - source_kind
      - parse_status
      - created_at
      - updated_at
      title: DocumentDetailRead
    DocumentRead:
      properties:
        id:
          type: integer
          title: Id
        source_path:
          type: string
          title: Source Path
        source_file_name:
          type: string
          title: Source File Name
        file_type:
          type: string
          title: File Type
        source_kind:
          type: string
          title: Source Kind
        parse_status:
          type: string
          title: Parse Status
        chunk_count:
          type: integer
          title: Chunk Count
          default: 0
        last_indexed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Indexed At
      type: object
      required:
      - id
      - source_path
      - source_file_name
      - file_type
      - source_kind
      - parse_status
      title: DocumentRead
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
    OpsOverviewResponse:
      properties:
        knowledge_bases:
          type: integer
          title: Knowledge Bases
          default: 0
        documents:
          type: integer
          title: Documents
          default: 0
        chunks:
          type: integer
          title: Chunks
          default: 0
        tasks_total:
          type: integer
          title: Tasks Total
          default: 0
        tasks_pending:
          type: integer
          title: Tasks Pending
          default: 0
        tasks_running:
          type: integer
          title: Tasks Running
          default: 0
        tasks_claimed_stale:
          type: integer
          title: Tasks Claimed Stale
          default: 0
        avg_task_wait_ms:
          type: integer
          title: Avg Task Wait Ms
          default: 0
        avg_task_run_ms:
          type: integer
          title: Avg Task Run Ms
          default: 0
        long_term_memories:
          type: integer
          title: Long Term Memories
          default: 0
        short_term_memories:
          type: integer
          title: Short Term Memories
          default: 0
        memory_ingestions:
          type: integer
          title: Memory Ingestions
          default: 0
        retrieval_logs:
          type: integer
          title: Retrieval Logs
          default: 0
        source_assets_missing:
          type: integer
          title: Source Assets Missing
          default: 0
        source_assets_stale:
          type: integer
          title: Source Assets Stale
          default: 0
        uploads:
          type: integer
          title: Uploads
          default: 0
      type: object
      title: OpsOverviewResponse
...
    TaskFailureRead:
      properties:
        id:
          type: integer
          title: Id
        kb_id:
          type: integer
          title: Kb Id
        task_type:
          type: string
          title: Task Type
        status:
          type: string
          title: Status
        trace_id:
          type: string
          title: Trace Id
          default: ''
        source_paths:
          items:
            type: string
          type: array
          title: Source Paths
        error_message:
          type: string
          title: Error Message
          default: ''
        stats_json:
          type: object
          title: Stats Json
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - kb_id
      - task_type
      - status
      - created_at
      title: TaskFailureRead
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WorkerStatusRead:
      properties:
        worker_name:
          type: string
          title: Worker Name
        status:
          type: string
          title: Status
        last_seen_at:
          type: string
          format: date-time
          title: Last Seen At
        last_processed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Processed At
        processed_count:
          type: integer
          title: Processed Count
          default: 0
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
        active_tasks_count:
          type: integer
          title: Active Tasks Count
          default: 0
      type: object
      required:
      - worker_name
      - status
      - last_seen_at
      title: WorkerStatusRead
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://127.0.0.1:8000
  description: 本地开发环境