# knowledge/api/routes_ops.py
from __future__ import annotations

from datetime import timedelta

from fastapi import APIRouter, Depends
from sqlalchemy import func, select
from sqlalchemy.orm import Session

from knowledge.api.deps import get_current_wallet
from knowledge.db.session import get_db
from knowledge.models import (
    ImportedChunk,
    ImportedDocument,
    ImportTask,
    KnowledgeBase,
    LongTermMemory,
    MemoryIngestionEvent,
    RetrievalLog,
    ShortTermMemory,
    SourceAsset,
    UploadRecord,
    WorkerStatus,
)
from knowledge.core.settings import get_settings
from knowledge.schemas.ops import OpsOverviewResponse, StoresHealthResponse, TaskFailureRead, WorkerStatusRead
from knowledge.services.vector_store import build_vector_store
from knowledge.services.task_queue import TaskQueueService
from knowledge.utils.time import utc_now


router = APIRouter(prefix="/ops", tags=["ops"], dependencies=[Depends(get_current_wallet)])
task_queue_service = TaskQueueService()


@router.get("/overview", response_model=OpsOverviewResponse)
def overview(db: Session = Depends(get_db)) -> dict:
    settings = get_settings()
    stale_before = utc_now() - timedelta(seconds=max(15, int(settings.worker_run_lease_ttl_seconds)))
    running_tasks = int(
        db.scalar(select(func.count(ImportTask.id)).where(ImportTask.status.in_(("running", "cancel_requested")))) or 0
    )
    stale_claims = int(
        db.scalar(
            select(func.count(ImportTask.id))
            .where(ImportTask.status.in_(("running", "cancel_requested")))
            .where(
                (ImportTask.heartbeat_at.is_(None) & (ImportTask.started_at < stale_before))
                | (ImportTask.heartbeat_at < stale_before)
            )
        )
        or 0
    )
    completed_tasks = list(
        db.scalars(
            select(ImportTask)
            .where(ImportTask.finished_at.is_not(None))
            .order_by(ImportTask.finished_at.desc(), ImportTask.id.desc())
            .limit(50)
        ).all()
    )
    wait_samples = [
        max(0, int((task.started_at - task.created_at).total_seconds() * 1000))
        for task in completed_tasks
        if task.started_at is not None
    ]
    run_samples = [
        max(0, int((task.finished_at - task.started_at).total_seconds() * 1000))
        for task in completed_tasks
        if task.started_at is not None and task.finished_at is not None
    ]
    return {
        "knowledge_bases": int(db.scalar(select(func.count(KnowledgeBase.id))) or 0),
        "documents": int(db.scalar(select(func.count(ImportedDocument.id))) or 0),
        "chunks": int(db.scalar(select(func.count(ImportedChunk.id))) or 0),
        "tasks_total": int(db.scalar(select(func.count(ImportTask.id))) or 0),
        "tasks_pending": int(db.scalar(select(func.count(ImportTask.id)).where(ImportTask.status == "pending")) or 0),
        "tasks_running": running_tasks,
        "tasks_claimed_stale": stale_claims,
        "avg_task_wait_ms": int(sum(wait_samples) / len(wait_samples)) if wait_samples else 0,
        "avg_task_run_ms": int(sum(run_samples) / len(run_samples)) if run_samples else 0,
        "long_term_memories": int(db.scalar(select(func.count(LongTermMemory.id))) or 0),
        "short_term_memories": int(db.scalar(select(func.count(ShortTermMemory.id))) or 0),
        "memory_ingestions": int(db.scalar(select(func.count(MemoryIngestionEvent.id))) or 0),
        "retrieval_logs": int(db.scalar(select(func.count(RetrievalLog.id))) or 0),
        "source_assets_missing": int(
            db.scalar(select(func.count(SourceAsset.id)).where(SourceAsset.availability_status == "missing")) or 0
        ),
        "source_assets_stale": int(
            db.scalar(select(func.count(SourceAsset.id)).where(SourceAsset.availability_status == "changed")) or 0
        ),
        "uploads": int(db.scalar(select(func.count(UploadRecord.id))) or 0),
    }


@router.get("/stores/health", response_model=StoresHealthResponse)
def stores_health(db: Session = Depends(get_db)) -> dict:
    settings = get_settings()
    try:
        db.execute(select(1))
        db_status = "ok"
    except Exception as exc:  # noqa: BLE001
        db_status = f"error: {exc}"
    vector_backend = build_vector_store()
    try:
        vector_status = vector_backend.health()
    except Exception as exc:  # noqa: BLE001
        vector_status = {"backend": settings.vector_store_mode, "status": f"error: {exc}"}
    model_status = "configured" if settings.model_provider_mode != "mock" and settings.model_gateway_base_url else "mock-or-not-configured"
    return {
        "database": db_status,
        "vector_store_mode": settings.vector_store_mode,
        "vector_store_status": vector_status,
        "model_provider_mode": settings.model_provider_mode,
        "model_provider_status": model_status,
        "object_storage_endpoint": settings.object_storage_endpoint,
        "object_storage_region": settings.object_storage_region,
        # Deprecated response fields retained for old console clients.
        "warehouse_gateway_mode": settings.warehouse_gateway_mode,
        "warehouse_base_url": settings.warehouse_base_url,
    }


@router.get("/workers", response_model=list[WorkerStatusRead])
def workers(db: Session = Depends(get_db)) -> list[dict]:
    rows = list(db.scalars(select(WorkerStatus).order_by(WorkerStatus.worker_name.asc())).all())
    now = utc_now()
    stale_before = task_queue_service.stale_before()
    active_by_worker = task_queue_service.active_tasks_by_worker(db, stale_before)
    results = []
    for row in rows:
        stale = row.last_seen_at < now - timedelta(seconds=90)
        results.append(
            {
                "worker_name": row.worker_name,
                "status": "stale" if stale and row.status != "error" else row.status,
                "last_seen_at": row.last_seen_at,
                "last_processed_at": row.last_processed_at,
                "processed_count": row.processed_count,
                "last_error": row.last_error,
                "active_tasks_count": active_by_worker.get(row.worker_name, 0),
            }
        )
    return results


@router.get("/tasks/failures", response_model=list[TaskFailureRead])
def recent_task_failures(trace_id: str | None = None, limit: int = 10, db: Session = Depends(get_db)) -> list[dict]:
    query_limit = max(limit, 100) if trace_id else limit
    rows = list(
        db.scalars(
            select(ImportTask)
            .where(ImportTask.status.in_(["failed", "partial_success"]))
            .order_by(ImportTask.finished_at.desc(), ImportTask.created_at.desc())
            .limit(query_limit)
        ).all()
    )
    if trace_id:
        rows = [
            row
            for row in rows
            if (row.stats_json or {}).get("trace_id") == trace_id
            or trace_id in ((row.stats_json or {}).get("related_trace_ids") or [])
        ]
        rows = rows[:limit]
    return [
        {
            "id": row.id,
            "kb_id": row.kb_id,
            "task_type": row.task_type,
            "status": row.status,
            "trace_id": (row.stats_json or {}).get("trace_id", ""),
            "source_paths": row.source_paths,
            "error_message": row.error_message,
            "stats_json": row.stats_json,
            "finished_at": row.finished_at,
            "created_at": row.created_at,
        }
        for row in rows
    ]